requirements.txt
pyspacemouse>=1.1.3
pynput>=1.7.6
PyQt6>=6.4.0
pyinstaller>=5.6.2
pytest>=7.3.1
keyboard>=0.13.5
pydirectinput-rgx>=2.1.2
psutil>=5.9.0 